\name{emMDR-package}
\alias{emMDR-package}
\alias{emMDR}
\docType{package}
\title{
The Multifactor Dimensionality Reduction (MDR) Analysis for Imcomplete Data
}
\description{
This provides various approaches to handling missing values for the MDR analysis of incomplete data
}
\details{
\tabular{ll}{
Package: \tab emMDR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2008-09-21\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}

emmdr(dataset, colresp, cs, combi, cv.fold = 10, na.method = 0, randomize = FALSE)
}
\author{
Junghyun Namkung, Taeyoung Hwang, MinSeok Kwon, Sunggon Yi and Wonil Chung
Maintainer: Junghyun Namkung <jh.namkung@gmail.com>
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{

}
