\name{imputeMDR-package}
\alias{imputeMDR-package}
\alias{imputeMDR}
\docType{package}
\title{
The Multifactor Dimensionality Reduction (MDR) Analysis for Imcomplete Data
}
\description{
This provides various approaches to handling missing values for the MDR analysis of incomplete data to identify gene-gene interactions using biallelic marker data in genetic association studies
}
\details{
\tabular{ll}{
Package: \tab imputeMDR\cr
Type: \tab Package\cr
Version: \tab 1.1.1\cr
Date: \tab 2011-06-28\cr
License: \tab GPL (>2) \cr
LazyLoad: \tab yes\cr
}

impute.mdr()
}
\author{
Junghyun Namkung, Taeyoung Hwang, MinSeok Kwon, Sunggon Yi and Wonil Chung
Maintainer: Junghyun Namkung <jh.namkung@gmail.com>
}
\references{
Namkung J, Elston RC, Yang JM, Park T. "Identification of gene-gene interactions in the presence of missing data using the multifactor dimensionality reduction method" Genet Epidemiol. 2009 Nov;33(7):646-56.
}
\keyword{ package }
\seealso{
\code{\link[imputeMDR]{impute.mdr}}
}
\examples{
}
