  \name{incomplete}
  \docType{data}
  \alias{incomplete}
  \title{A simulated example containing missing values}
  \description{
    This data is a simulated case-control study data example with 10% missing values.
    10 SNPs for 200 cases/200 controls were included in this data.
    The first column is for the response variable.
  }
  \usage{incomplete}
  \format{A matrix of 400*11 dimension}
  \keyword{datasets}

 
